// src/types/types.ts
export interface HotDealDataType {
  category: string | null;
  hit: number;
  is_soldout: boolean;
  num_reply: number;
  site: string | undefined;
  site_id: string;
  thumbnail_url: string | null;
  title: string;
  updated_at: Date;
  uploaded_time: Date;
  url: string | undefined;
}

// export interface CategoryKeywordType {
//   category: string;
//   keyword: string;
// }